+++ porcupine-core/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased changes
+
+- Depend on monad-bayes by default since it is now on Hackage
+
 # Ver 0.1.0.1
 
 - Added a flag in porcupine-core/package.yaml so we don't depend on monad-bayes